+++ src/transform/regress_out/test.py
@@ -10,32 +10,50 @@ meta = {
 }
 ## VIASH END
 
-orig_input = f"{meta['resources_dir']}/pbmc_1k_protein_v3/pbmc_1k_protein_v3_filtered_feature_bc_matrix.h5mu"
+@pytest.fixture
+def input_path():
+    return f"{meta['resources_dir']}/pbmc_1k_protein_v3/pbmc_1k_protein_v3_filtered_feature_bc_matrix.h5mu"
 
-def test_regress_out(run_component, tmp_path):
-    input = tmp_path / "input.h5mu"
-    output = tmp_path / "output.h5mu"
 
-    # add column to obs
-    mu_orig_input = mu.read_h5mu(orig_input)
-    mu_orig_input.obs["var"] = np.random.rand(mu_orig_input.n_obs)
-    mu_orig_input.mod["rna"].obs["var"] = mu_orig_input.obs["var"]
-    mu_orig_input.mod["prot"].obs["var"] = mu_orig_input.obs["var"]
-    mu_orig_input.write_h5mu(input)
+@pytest.fixture
+def input_data(input_path):
+    return mu.read_h5mu(input_path)
+
+
+@pytest.fixture
+def input_h5mu(input_data):
+    input_data.obs["var"] = np.random.rand(input_data.n_obs)
+    input_data.mod["rna"].obs["var"] = input_data.obs["var"]
+    input_data.mod["prot"].obs["var"] = input_data.obs["var"]
+    input_data.mod["rna"].layers["input"] = input_data.mod["rna"].X
+    return input_data
+
+
+@pytest.fixture
+def input_h5mu_path(write_mudata_to_file, input_h5mu):
+    return write_mudata_to_file(input_h5mu)
+
+
+@pytest.fixture
+def output_h5mu_path(tmp_path):
+    return tmp_path / "output.h5mu"
+
+
+def test_regress_out(run_component, input_h5mu_path, output_h5mu_path):
 
     # execute command
     cmd_pars = [
-        "--input", input,
-        "--output", str(output),
+        "--input", input_h5mu_path,
+        "--output", output_h5mu_path,
         "--obs_keys", "var",
         "--output_compression", "gzip"
     ]
     run_component(cmd_pars)
 
-    assert output.is_file(), "No output was created."
+    assert output_h5mu_path.is_file(), "No output was created."
 
-    mu_input = mu.read_h5mu(input)
-    mu_output = mu.read_h5mu(output)
+    mu_input = mu.read_h5mu(input_h5mu_path)
+    mu_output = mu.read_h5mu(output_h5mu_path)
 
     assert "rna" in mu_output.mod, 'Output should contain data.mod["prot"].'
     assert "prot" in mu_output.mod, 'Output should contain data.mod["prot"].'
@@ -49,6 +67,29 @@ def test_regress_out(run_component, tmp_path):
     assert prot_in.shape == prot_out.shape, "Should have same shape as before"
 
     assert np.mean(rna_in.X) != np.mean(rna_out.X), "Expression should have changed"
+
+
+def test_regress_out_with_layers(run_component, input_h5mu_path, output_h5mu_path):
+
+    # execute command
+    cmd_pars = [
+        "--input", input_h5mu_path,
+        "--output", output_h5mu_path,
+        "--obs_keys", "var",
+        "--input_layer", "input",
+        "--output_layer", "output",
+        "--output_compression", "gzip"
+    ]
+    run_component(cmd_pars)
+
+    mu_input = mu.read_h5mu(input_h5mu_path)
+    mu_output = mu.read_h5mu(output_h5mu_path)
+
+    rna_in = mu_input.mod["rna"]
+    rna_out = mu_output.mod["rna"]
+
+    assert np.mean(rna_in.layers["input"]) != np.mean(rna_out.layers["output"]), "RNA expression should have changed"
+
 
 if __name__ == '__main__':
     sys.exit(pytest.main([__file__]))